clean up responsibilities
