+++ mod/cli/pkg/builder/builder.go
@@ -28,38 +28,35 @@ import (
 	"cosmossdk.io/log"
 	cmdlib "github.com/berachain/beacon-kit/mod/cli/pkg/commands"
 	"github.com/berachain/beacon-kit/mod/primitives"
-	cmtcfg "github.com/cometbft/cometbft/config"
 	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/config"
 	servertypes "github.com/cosmos/cosmos-sdk/server/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
-// CLIBuilder is a builder that incrementally constructs a CLI command.
+// CLIBuilder is the builder for the commands.Root (root command).
 type CLIBuilder[T servertypes.Application] struct {
 	depInjectCfg depinject.Config
 	name         string
 	description  string
-	components   []any
-	moduleDeps   []any
-	runHandler   func(cmd *cobra.Command,
-		customAppConfigTemplate string,
-		customAppConfig interface{},
-		cmtConfig *cmtcfg.Config,
-	) error
-	appCreator   servertypes.AppCreator[T]
-	rootCmdSetup func(cmd *cmdlib.Root,
-		mm *module.Manager,
-		appCreator servertypes.AppCreator[T],
-		chainSpec primitives.ChainSpec,
-	)
+	// components is a list of component providers for depinject.
+	components []any
+	// supplies is a list of suppliers for depinject.
+	supplies   []any
+	runHandler runHandler
+	// appCreator is a function that builds the Node, eventually called by the
+	// cosmos-sdk.
+	appCreator servertypes.AppCreator[T]
+	// rootCmdSetup is a function that sets up the root command.
+	rootCmdSetup rootCmdSetup[T]
 }
 
 // New returns a new CLIBuilder with the given options.
 func New[T servertypes.Application](opts ...Opt[T]) *CLIBuilder[T] {
-	cb := &CLIBuilder[T]{}
+	cb := &CLIBuilder[T]{
+		supplies: []any{log.NewLogger(os.Stdout), viper.GetViper()},
+	}
 	for _, opt := range opts {
 		opt(cb)
 	}
@@ -68,7 +65,7 @@ func New[T servertypes.Application](opts ...Opt[T]) *CLIBuilder[T] {
 
 // Build builds the CLI commands.
 func (cb *CLIBuilder[T]) Build() (*cmdlib.Root, error) {
-	// dependencies for the root command
+	// allocate memory to hold the dependencies
 	var (
 		autoCliOpts autocli.AppOptions
 		mm          *module.Manager
@@ -80,11 +77,7 @@ func (cb *CLIBuilder[T]) Build() (*cmdlib.Root, error) {
 		depinject.Configs(
 			cb.depInjectCfg,
 			depinject.Supply(
-				append([]any{
-					log.NewLogger(os.Stdout),
-					viper.GetViper(),
-				},
-					cb.moduleDeps...)...,
+				cb.supplies...,
 			),
 			depinject.Provide(
 				cb.components...,
@@ -98,66 +91,37 @@ func (cb *CLIBuilder[T]) Build() (*cmdlib.Root, error) {
 		return nil, err
 	}
 
-	cmd := &cobra.Command{
-		Use:   cb.name,
-		Short: cb.description,
-		PersistentPreRunE: func(cmd *cobra.Command, _ []string) error {
-			// set the default command outputs
-			cmd.SetOut(cmd.OutOrStdout())
-			cmd.SetErr(cmd.ErrOrStderr())
+	// pass in deps to build the root command
+	rootCmd := cmdlib.New(
+		cb.name,
+		cb.description,
+		defaultRunHandler(cb.runHandler),
+		clientCtx,
+	)
 
-			var err error
-			clientCtx, err = client.ReadPersistentCommandFlags(
-				clientCtx,
-				cmd.Flags(),
-			)
-			if err != nil {
-				return err
-			}
-
-			customClientTemplate, customClientConfig := InitClientConfig()
-			clientCtx, err = config.CreateClientConfig(
-				clientCtx,
-				customClientTemplate,
-				customClientConfig,
-			)
-			if err != nil {
-				return err
-			}
-
-			if err = client.SetCmdClientContextHandler(
-				clientCtx, cmd,
-			); err != nil {
-				return err
-			}
-
-			return cb.runHandler(
-				cmd,
-				DefaultAppConfigTemplate(),
-				DefaultAppConfig(),
-				DefaultCometConfig(),
-			)
-		},
-	}
-
+	// apply default root command setup
 	cmdlib.DefaultRootCommandSetup(
-		cmdlib.New(cmd),
+		rootCmd,
 		mm,
 		cb.appCreator,
 		chainSpec,
 	)
 
-	if err := autoCliOpts.EnhanceRootCommand(cmd); err != nil {
+	// enhance the root command with the autoCliOpts
+	if err := rootCmd.Enhance(autoCliOpts.EnhanceRootCommand); err != nil {
 		return nil, err
 	}
 
-	return cmdlib.New(cmd), nil
+	return rootCmd, nil
 }
 
-// InitClientConfig sets up the default client configuration, allowing for
-// overrides.
-func InitClientConfig() (string, interface{}) {
-	clientCfg := config.DefaultConfig()
-	clientCfg.KeyringBackend = "test"
-	return config.DefaultClientConfigTemplate, clientCfg
+// defaultRunHandler returns a runHandler that uses the default configuration.
+func defaultRunHandler(runHandler runHandler) func(cmd *cobra.Command) error {
+	return func(cmd *cobra.Command) error {
+		return runHandler(cmd,
+			DefaultAppConfigTemplate(),
+			DefaultAppConfig(),
+			DefaultCometConfig(),
+		)
+	}
 }
