+++ mod/cli/pkg/commands/root.go
@@ -21,23 +21,9 @@
 package commands
 
 import (
-	confixcmd "cosmossdk.io/tools/confix/cmd"
-	"github.com/berachain/beacon-kit/mod/cli/pkg/commands/client"
-	"github.com/berachain/beacon-kit/mod/cli/pkg/commands/cometbft"
-	"github.com/berachain/beacon-kit/mod/cli/pkg/commands/deposit"
-	"github.com/berachain/beacon-kit/mod/cli/pkg/commands/genesis"
-	"github.com/berachain/beacon-kit/mod/cli/pkg/commands/jwt"
-	"github.com/berachain/beacon-kit/mod/cli/pkg/flags"
-	"github.com/berachain/beacon-kit/mod/primitives"
-	"github.com/cosmos/cosmos-sdk/client/keys"
-	"github.com/cosmos/cosmos-sdk/client/pruning"
-	"github.com/cosmos/cosmos-sdk/client/snapshot"
-	"github.com/cosmos/cosmos-sdk/server"
+	sdkclient "github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/config"
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
-	servertypes "github.com/cosmos/cosmos-sdk/server/types"
-	"github.com/cosmos/cosmos-sdk/types/module"
-	"github.com/cosmos/cosmos-sdk/version"
-	genutilcli "github.com/cosmos/cosmos-sdk/x/genutil/client/cli"
 	"github.com/spf13/cobra"
 )
 
@@ -46,61 +32,61 @@ type Root struct {
 	cmd *cobra.Command
 }
 
-// New returns a pointer to a new Root.
-func New(rootcmd *cobra.Command) *Root {
+// New returns a new root command with the provided configuration.
+func New(name string,
+	description string,
+	runHandler runHandler,
+	clientCtx sdkclient.Context,
+) *Root {
+	cmd := &cobra.Command{
+		Use:   name,
+		Short: description,
+		PersistentPreRunE: func(cmd *cobra.Command, _ []string) error {
+			// set the default command outputs
+			cmd.SetOut(cmd.OutOrStdout())
+			cmd.SetErr(cmd.ErrOrStderr())
+
+			var err error
+			clientCtx, err = sdkclient.ReadPersistentCommandFlags(
+				clientCtx,
+				cmd.Flags(),
+			)
+			if err != nil {
+				return err
+			}
+
+			customClientTemplate, customClientConfig := InitClientConfig()
+			clientCtx, err = config.CreateClientConfig(
+				clientCtx,
+				customClientTemplate,
+				customClientConfig,
+			)
+			if err != nil {
+				return err
+			}
+
+			if err = sdkclient.SetCmdClientContextHandler(
+				clientCtx, cmd,
+			); err != nil {
+				return err
+			}
+
+			return runHandler(cmd)
+		},
+	}
 	return &Root{
-		cmd: rootcmd,
+		cmd: cmd,
 	}
 }
 
 // Run executes the root command.
-func (root Root) Run(defaultNodeHome string) error {
+func (root *Root) Run(defaultNodeHome string) error {
 	return svrcmd.Execute(
 		root.cmd, "", defaultNodeHome,
 	)
 }
 
-// DefaultRootCommandSetup sets up the default commands for the root command.
-func DefaultRootCommandSetup[T servertypes.Application](
-	root *Root,
-	mm *module.Manager,
-	appCreator servertypes.AppCreator[T],
-	chainSpec primitives.ChainSpec,
-) {
-	// Setup the custom start command options.
-	startCmdOptions := server.StartCmdOptions[T]{
-		AddFlags: flags.AddBeaconKitFlags,
-	}
-
-	// Add all the commands to the root command.
-	root.cmd.AddCommand(
-		// `comet`
-		cometbft.Commands(appCreator),
-		// `client`
-		client.Commands[T](),
-		// `config`
-		confixcmd.ConfigCommand(),
-		// `init`
-		genutilcli.InitCmd(mm),
-		// `genesis`
-		genesis.Commands(chainSpec),
-		// `deposit`
-		deposit.Commands(chainSpec),
-		// `jwt`
-		jwt.Commands(),
-		// `keys`
-		keys.Commands(),
-		// `prune`
-		pruning.Cmd(appCreator),
-		// `rollback`
-		server.NewRollbackCmd(appCreator),
-		// `snapshots`
-		snapshot.Cmd(appCreator),
-		// `start`
-		server.StartCmdWithOptions(appCreator, startCmdOptions),
-		// `status`
-		server.StatusCommand(),
-		// `version`
-		version.NewVersionCommand(),
-	)
+// Enhance applies the given enhancer to the root command.
+func (root *Root) Enhance(enhance enhancer) error {
+	return enhance(root.cmd)
 }
